script_ad_spend/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
google-cloud-bigquery>=3.25.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
authlib>=1.3.0
pycryptodome>=3.20.0

